+++ client/src/App.tsx
@@ -1,5 +1,5 @@
-// import HeaderBefore from "./components/HeaderBefore";
-import HeaderAfter from "./components/HeaderAfter";
+// import HeaderBefore from "./components/Header/HeaderBefore";
+import HeaderAfter from "./components/Header/HeaderAfter";
 
 function App() {
   return (
